+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix removal of monthly filter dropdown on exhibitions_space
+- View link in row quick actions ([#6](https://github.com/mdibella-dev/congressomat/issues/6))
 
 <br>